requirements.txt
numpy
scipy
torch>=1.7
torchvision
tensorboard
hydra-core
optuna
hydra-optuna-sweeper
tqdm
timm
